pyproject.toml
[project]
name = "autodrive-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "eventlet==0.33.3",
    "flask==1.1.1",
    "flask-socketio==4.1.0",
    "gevent-websocket==0.10.1",
    "itsdangerous==2.0.1",
    "jinja2==3.0.3",
    "numpy>=2.2.6",
    "opencv-contrib-python>=4.11.0.86",
    "pillow>=11.3.0",
    "python-engineio==3.13.0",
    "python-socketio==4.2.0",
    "werkzeug==2.0.3",
]

[tool.uv]
package = true
